Moving AllDataBuilder in interests of hierarchy
